+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/query/plan/bitmap/ComposedBitmapIndexAggregate.java
@@ -44,6 +44,7 @@ import com.apple.foundationdb.record.query.expressions.QueryKeyExpressionWithCom
 import com.apple.foundationdb.record.query.plan.RecordQueryPlanner;
 import com.apple.foundationdb.record.query.plan.planning.FilterSatisfiedMask;
 import com.apple.foundationdb.record.query.plan.plans.RecordQueryCoveringIndexPlan;
+import com.apple.foundationdb.record.query.plan.plans.RecordQueryPlan;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
@@ -82,9 +83,9 @@ public class ComposedBitmapIndexAggregate {
      * @return an {@code Optional} query plan or {@code Optional.empty} if planning is not possible
      */
     @Nonnull
-    public static Optional<ComposedBitmapIndexQueryPlan> tryPlan(@Nonnull RecordQueryPlanner planner,
-                                                                 @Nonnull RecordQuery query,
-                                                                 @Nonnull IndexAggregateFunction indexAggregateFunction) {
+    public static Optional<RecordQueryPlan> tryPlan(@Nonnull RecordQueryPlanner planner,
+                                                    @Nonnull RecordQuery query,
+                                                    @Nonnull IndexAggregateFunction indexAggregateFunction) {
         if (query.getFilter() == null || query.getSort() != null) {
             return Optional.empty();
         }
@@ -99,12 +100,18 @@ public class ComposedBitmapIndexAggregate {
      * @return an {@code Optional} query plan or {@code Optional.empty} if planning is not possible
      */
     @Nonnull
-    public Optional<ComposedBitmapIndexQueryPlan> tryPlan(@Nonnull RecordQueryPlanner planner,
-                                                          @Nonnull RecordQuery.Builder queryBuilder) {
+    public Optional<RecordQueryPlan> tryPlan(@Nonnull RecordQueryPlanner planner,
+                                             @Nonnull RecordQuery.Builder queryBuilder) {
         final List<RecordQueryCoveringIndexPlan> indexScans = new ArrayList<>();
         final Map<IndexNode, ComposedBitmapIndexQueryPlan.IndexComposer> indexComposers = new IdentityHashMap<>();
-        return Optional.ofNullable(plan(root, queryBuilder, planner, indexScans, indexComposers))
-                .map(composer -> new ComposedBitmapIndexQueryPlan(indexScans, composer));
+        final ComposedBitmapIndexQueryPlan.ComposerBase composer = plan(root, queryBuilder, planner, indexScans, indexComposers);
+        if (composer == null || indexScans.isEmpty()) {
+            return Optional.empty();
+        }
+        if (indexScans.size() == 1) {
+            return Optional.ofNullable(indexScans.get(0));
+        }
+        return Optional.of(new ComposedBitmapIndexQueryPlan(indexScans, composer));
     }
 
     /**